Timer
Handy work timer that manages time in intervals of 25 minute work sessions and 5 minute breaks. once that cycle has repeated 4 times, take a 20 minute break. timer indicates the status it is in, the number of repetitions completed, and pops to front of screen when the state of the timer updates